+++ examples/Hex-String/main.go
@@ -6,10 +6,10 @@ package main
 
 import (
 	"fmt"
+	g "github.com/Rakortx/gosnmp"
 	"log"
-	"strings"
 	"strconv"
-	g "github.com/gosnmp/gosnmp"
+	"strings"
 )
 
 func main() {
@@ -36,7 +36,7 @@ func main() {
 			value := variable.Value.([]byte)
 			if strings.Contains(strconv.Quote(string(value)), "\\x") {
 				tmp := ""
-                for i := 0; i < len(value); i++ {
+				for i := 0; i < len(value); i++ {
 					tmp += fmt.Sprintf("%v", value[i])
 					if i != (len(value) - 1) {
 						tmp += " "