+++ examples/example/main.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 	"log"
 
-	g "github.com/gosnmp/gosnmp"
+	g "github.com/Rakortx/gosnmp"
 )
 
 func main() {