+++ examples/example2/main.go
@@ -49,7 +49,7 @@ import (
 	"strconv"
 	"time"
 
-	g "github.com/gosnmp/gosnmp"
+	g "github.com/Rakortx/gosnmp"
 )
 
 func main() {