+++ examples/linux_vrf/main.go
@@ -10,7 +10,7 @@ import (
 	"syscall"
 	"time"
 
-	g "github.com/gosnmp/gosnmp"
+	g "github.com/Rakortx/gosnmp"
 )
 
 func main() {