+++ examples/trapsend_v1/main.go
@@ -7,7 +7,7 @@ package main
 import (
 	"log"
 
-	g "github.com/gosnmp/gosnmp"
+	g "github.com/Rakortx/gosnmp"
 )
 
 func main() {