+++ gosnmp_api_test.go
@@ -19,7 +19,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/gosnmp/gosnmp"
+	"github.com/Rakortx/gosnmp"
 )
 
 func TestAPIConfigTypes(t *testing.T) {